+++ tests/Console/ModelsCommand/ModelHooks/Test.php
@@ -8,6 +8,7 @@ use Barryvdh\LaravelIdeHelper\Console\ModelsCommand;
 use Barryvdh\LaravelIdeHelper\Tests\Console\ModelsCommand\AbstractModelsCommand;
 use Barryvdh\LaravelIdeHelper\Tests\Console\ModelsCommand\ModelHooks\Hooks\CustomMethod;
 use Barryvdh\LaravelIdeHelper\Tests\Console\ModelsCommand\ModelHooks\Hooks\CustomProperty;
+use Barryvdh\LaravelIdeHelper\Tests\Console\ModelsCommand\ModelHooks\Hooks\CustomTags;
 use Barryvdh\LaravelIdeHelper\Tests\Console\ModelsCommand\ModelHooks\Hooks\UnsetMethod;
 use Illuminate\Filesystem\Filesystem;
 use Mockery;
@@ -27,6 +28,7 @@ class Test extends AbstractModelsCommand
             'model_hooks' => [
                 CustomProperty::class,
                 CustomMethod::class,
+                CustomTags::class,
                 UnsetMethod::class,
             ],
         ]);
@@ -80,6 +82,7 @@ use Illuminate\Database\Eloquent\Model;
  * @method static \Illuminate\Database\Eloquent\Builder|Simple query()
  * @method static \Illuminate\Database\Eloquent\Builder|Simple whereId($value)
  * @mixin \Eloquent
+ * @phpstorm-ignore argument.type
  */
 class Simple extends Model
 {